refactor(webkit): move SegmentedButton from wip to actions layer
Relocate the component under actions, add the canonical actions export, and update Storybook imports to @aziontech/webkit/actions/segmented-button.
