+++ apps/storybook/src/stories/webkit/actions/segmented-button/SegmentedButton.stories.js
@@ -1,4 +1,4 @@
-import SegmentedButton from '@aziontech/webkit/segmented-button'
+import SegmentedButton from '@aziontech/webkit/actions/segmented-button'
 import { ref, watch } from 'vue'
 
 const defaultOptions = [
