+++ apps/storybook/src/stories/webkit/templates/ChangePlanDrawer.stories.js
@@ -9,7 +9,7 @@ import DrawerTitle from '@aziontech/webkit/overlay/drawer-title'
 import DrawerTrigger from '@aziontech/webkit/overlay/drawer-trigger'
 import PanelContent from '@aziontech/webkit/overlay/panel-content'
 import PanelHeader from '@aziontech/webkit/overlay/panel-header'
-import SegmentedButton from '@aziontech/webkit/segmented-button'
+import SegmentedButton from '@aziontech/webkit/actions/segmented-button'
 import { expect, userEvent, within } from '@storybook/test'
 import { ref } from 'vue'
